+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.11.2...HEAD) - TBD
 
+### :bug: Fixed
+
+- False positive `negative_data_rejection` for path parameters with `format: uuid` and a `pattern` constraint in the fuzzing phase captured valid resource IDs are used. [#3603](https://github.com/schemathesis/schemathesis/issues/3603)
+
 ## [4.11.2](https://github.com/schemathesis/schemathesis/compare/v4.11.1...v4.11.2) - 2026-03-06
 
 ### :rocket: Added